+++ src/components/archive/card.tsx
@@ -1,6 +1,6 @@
-'use client';
+"use client";
 
-import { Eye, FileText } from 'lucide-react';
+import { Eye, FileText } from "lucide-react";
 
 interface ArchiveCardProps {
   id: string;
@@ -40,7 +40,7 @@ export default function ArchiveCard({
       </div>
 
       {/* 콘텐츠 영역 */}
-      <div className="p-4 flex flex-col gap-3 flex-1 flex-grow">
+      <div className="p-4 flex flex-col gap-3 flex-1 grow">
         {/* 제목 */}
         <h3 className="text-sm font-bold text-gray-900 line-clamp-2 leading-snug">
           {title}
@@ -55,7 +55,7 @@ export default function ArchiveCard({
           </span>
           <span className="text-xs text-gray-400">{uploadedAt}</span>
         </div>
-        <div className="flex gap-2 text-xs text-gray-500 flex-shrink-0">
+        <div className="flex gap-2 text-xs text-gray-500 shrink-0">
           {views > 0 && (
             <span className="flex items-center gap-1">
               <Eye size={16} />
